product_classifier_tk/requirements.txt
# Core dependencies
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0

# AI/ML
ultralytics>=8.0.0

# Hardware (optional - only for Raspberry Pi)
pyserial>=3.5

# Note: RPi.GPIO is pre-installed on Raspberry Pi OS
# On Windows/Mac, the app will run without GPIO functionality

